+++ dev/examples/Drag-to-reorder.tsx
@@ -5,56 +5,29 @@ import { Reorder, useMotionValue, animate } from "framer-motion"
 const inactiveShadow = "0px 0px 0px rgba(0,0,0,0.8)"
 
 const Item = ({ item, axis }) => {
-    const x = useMotionValue(0)
     const y = useMotionValue(0)
-    const axisValue = axis === "y" ? y : x
     const boxShadow = useMotionValue(inactiveShadow)
-    // const dragControls = useDragControls()
-
-    useEffect(() => {
-        let isActive = false
-        axisValue.on("change", (latestY) => {
-            const wasActive = isActive
-            if (latestY !== 0) {
-                isActive = true
-                if (isActive !== wasActive) {
-                    animate(boxShadow, "5px 5px 10px rgba(0,0,0,0.3)")
-                }
-            } else {
-                isActive = false
-                if (isActive !== wasActive) {
-                    animate(boxShadow, inactiveShadow)
-                }
-            }
-        })
-    }, [y, boxShadow])
 
     return (
         <Reorder.Item
             value={item}
             id={item}
-            // dragListener={false}
-            // dragControls={dragControls}
             style={{ boxShadow, y }}
-            dragTransition={{ bounceStiffness: 2000, bounceDamping: 10000 }}
             transition={{ duration: 0.1 }}
         >
             <span>{item}</span>
-            <ReorderIcon
-            // dragControls={dragControls}
-            />
+            <ReorderIcon />
         </Reorder.Item>
     )
 }
 
 export const App = () => {
     const [items, setItems] = useState(initialItems)
-    const params = new URLSearchParams(window.location.search)
-    const axis = params.get("axis") || "y"
+    const axis = "y"
 
     return (
         <Reorder.Group
-            axis={axis}
+            axis="y"
             onReorder={setItems}
             style={axis === "y" ? verticalList : horizontalList}
             values={items}