+++ dev/tests/layout-shared-lightbox-crossfade.tsx
@@ -1,6 +1,11 @@
 import * as React from "react"
 import { CSSProperties, useState } from "react"
-import { motion, AnimatePresence, useIsPresent, Transition } from "framer-motion"
+import {
+    motion,
+    AnimatePresence,
+    useIsPresent,
+    Transition,
+} from "framer-motion"
 
 /**
  * This demonstrates children with layoutId animating