+++ packages/framer-motion/src/gestures/hover.ts
@@ -11,7 +11,7 @@ function addHoverEvent(node: VisualElement<Element>, isActive: boolean) {
     const callbackName = "onHover" + (isActive ? "Start" : "End")
 
     const handleEvent = (event: PointerEvent, info: EventInfo) => {
-        if (event.type === "touch" || isDragActive()) return
+        if (event.pointerType === "touch" || isDragActive()) return
 
         const props = node.getProps()