+++ packages/framer-motion/src/gestures/pan/PanSession.ts
@@ -93,12 +93,14 @@ interface PanSessionHandlers {
     onStart: PanHandler
     onMove: PanHandler
     onEnd: PanHandler
-    onSessionEnd: PanHandler
+    onSessionEnd: PanHandler,
+    resumeAnimation: () => void
 }
 
 interface PanSessionOptions {
-    transformPagePoint?: TransformPoint
+    transformPagePoint?: TransformPoint,
     contextWindow?: (Window & typeof globalThis) | null
+    dragSnapToOrigin?: boolean
 }
 
 interface TimestampedPoint extends Point {
@@ -145,6 +147,13 @@ export class PanSession {
     private removeListeners: Function
 
     /**
+     * For determining if an animation should resume after it is interupted
+     * 
+     * @internal
+     */
+    private dragSnapToOrigin: boolean
+    
+    /**
      * @internal
      */
     private contextWindow: PanSessionOptions["contextWindow"] = window
@@ -152,11 +161,12 @@ export class PanSession {
     constructor(
         event: PointerEvent,
         handlers: Partial<PanSessionHandlers>,
-        { transformPagePoint, contextWindow }: PanSessionOptions = {}
+        { transformPagePoint, contextWindow, dragSnapToOrigin = false }: PanSessionOptions = {}
     ) {
         // If we have more than one touch, don't start detecting this gesture
         if (!isPrimaryPointer(event)) return
 
+        this.dragSnapToOrigin = dragSnapToOrigin
         this.handlers = handlers
         this.transformPagePoint = transformPagePoint
         this.contextWindow = contextWindow || window
@@ -231,9 +241,11 @@ export class PanSession {
     private handlePointerUp = (event: PointerEvent, info: EventInfo) => {
         this.end()
 
+        const { onEnd, onSessionEnd, resumeAnimation } = this.handlers
+        
+        if(this.dragSnapToOrigin) resumeAnimation && resumeAnimation()
         if (!(this.lastMoveEvent && this.lastMoveEventInfo)) return
 
-        const { onEnd, onSessionEnd } = this.handlers
         const panInfo = getPanInfo(
             event.type === "pointercancel"
                 ? this.lastMoveEventInfo